requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-auth==0.9.5
django-allauth==0.57.0
drf-yasg==1.21.7
python-decouple==3.8
psycopg2-binary==2.9.9
requests==2.31.0
Pillow==10.1.0
channels==4.0.0
channels-redis==4.1.0